nested block priors

+++ _midynet/include/FastMIDyNet/types.h
@@ -28,7 +28,7 @@ typedef std::vector<BlockIndex> BlockSequence;
 typedef Matrix<size_t> EdgeMatrix;
 typedef CounterMap<std::pair<BlockIndex, size_t>> VertexLabeledDegreeCountsMap;
 typedef CounterMap<size_t> DegreeCountsMap;
-typedef size_t Level;
+typedef int Level;
 
 } // namespace FastMIDyNet
